treat argument as a float not an int in iso8601.py

+++ python/iso8601.py
@@ -5,6 +5,6 @@ import datetime, sys
 if 1 == len(sys.argv):
     epoch = datetime.datetime.utcnow()
 else:
-    epoch = datetime.datetime.fromtimestamp(int(sys.argv[1]))
+    epoch = datetime.datetime.fromtimestamp(float(sys.argv[1]))
 
 print(epoch.replace(microsecond=0).isoformat()+'Z')
